move function to extract rules IDs to utils package
